+++ client/verta/verta/operations/monitoring/notification_channels/__init__.py
@@ -1,0 +1,4 @@
+from ._notification_channel import (
+    _NotificationChannel,
+    SlackNotificationChannel,
+)
