+++ client/verta/verta/operations/monitoring/notification_channels/_entities/__init__.py
@@ -1,0 +1,1 @@
+from .notification_channel import NotificationChannel, NotificationChannels
