+++ client/verta/verta/operations/monitoring/notification_channels/_entities/notification_channel.py
@@ -2,8 +2,8 @@
 
 import warnings
 
-from ..._protos.public.monitoring import Alert_pb2 as _AlertService
-from ..._tracking import entity, _Context
+from ....._protos.public.monitoring import Alert_pb2 as _AlertService
+from ....._tracking import entity, _Context
 
 
 class NotificationChannel(entity._ModelDBEntity):
